add revision base class
